Implement most of statistics

+++ backend-dotnet/MMRProject.Api/DTOs/LeaderboardEntry.cs
@@ -4,7 +4,7 @@ public record LeaderboardEntry
 {
     public required long UserId { get; set; }
     public required string Name { get; set; }
-    public int? MMR { get; set; }
+    public long? MMR { get; set; }
     public int Wins { get; set; }
     public int Loses { get; set; }
     public int WinningStreak { get; set; }
